Move all unit tests to the tests folder
