+++ tests/unit/components/blocks/FilterOptions.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import { FilterOptions } from "./FilterOptions";
+import { FilterOptions } from "../../../../src/components/blocks/FilterOptions";
 
 describe("FilterOptions", () => {
   it("displays the correct filter options when corporaKey is present in themeConfig", () => {
